inference-service/requirements.txt
torch>=2.1.0
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic>=2.10.0
transformers>=4.49.0
accelerate>=0.33.0
sentencepiece==0.1.99
protobuf>=4.25.0
google-cloud-storage==2.14.0
google-cloud-firestore==2.14.0  # For API key authentication
requests==2.31.0
python-dotenv==1.0.0
bitsandbytes>=0.41.0  # For 8-bit quantization support (optional for local testing)
cachetools==5.3.2  # For in-memory API key caching
psutil>=5.9.0  # For system memory monitoring
peft>=0.13.0  # For LoRA adapter loading